fix(ci): 修复检查链路失败
CI 在修复 setup 后继续暴露出 clippy、host unit test 和 dependency audit 的阻断问题。本次把这些检查链路收敛到当前仓库真实支持的边界。

- 修复当前 riscv64/aarch64 clippy 下的 lint 失败，保留有所有权意义的字段并删除无用状态。

- 让 memory_types 只在裸机目标依赖 arch，host unit test 不再编译纯裸机 arch crate。

- 调整 cargo-deny 配置，使 workspace/path 依赖与 paste 传递依赖风险以显式策略处理。

验证：

- docker exec -w /workspace simplekernel-dev cargo fmt --all -- --check

- docker exec -w /workspace simplekernel-dev cargo clippy --target riscv64gc-unknown-none-elf -- -D warnings

- docker exec -w /workspace simplekernel-dev cargo clippy --target aarch64-unknown-none -- -D warnings

- docker exec -w /workspace simplekernel-dev cargo test -p memory_types -p config -p page_table_entry

- docker exec -w /workspace simplekernel-dev cargo deny check

- docker exec -w /workspace simplekernel-dev cargo xtask build --arch riscv64

- docker exec -w /workspace simplekernel-dev cargo xtask build --arch aarch64

- docker exec -w /workspace simplekernel-dev timeout 30 cargo xtask test --arch riscv64 --name heap-test --timeout 30

- git diff --check

+++ src/arch/riscv64/context.rs
@@ -162,9 +162,9 @@ impl CalleeSavedContext {
     /// - `s0` → 入口函数指针
     /// - `s1` → 入口函数参数
     pub fn init_for_kernel_thread(&mut self, kstack_top: usize, entry: fn(usize), arg: usize) {
-        self.ra = super::switch::kernel_thread_entry as unsafe extern "C" fn() as u64;
+        self.ra = super::switch::kernel_thread_entry as unsafe extern "C" fn() as usize as u64;
         self.sp = kstack_top as u64;
-        self.s0 = entry as u64;
+        self.s0 = entry as usize as u64;
         self.s1 = arg as u64;
     }
 }
